requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.1
click==8.1.7
cytoolz==0.12.3
dnspython==2.6.1
email_validator==2.1.1
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.2.3
eth-utils==4.1.0
eth_abi==5.1.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.2
frozenlist==1.4.1
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
orjson==3.10.3
parsimonious==0.10.0
protobuf==5.26.1
pycryptodome==3.20.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
rlp==4.0.1
rpds-py==0.18.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
toolz==0.12.1
typer==0.12.3
typing_extensions==4.11.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
web3==6.18.0
websockets==12.0
yarl==1.9.4
